elevate progress state to shared
