+++ src/app/features/auth/effects/auth.effects.spec.ts
@@ -9,10 +9,10 @@ import { AuthEffects } from './auth.effects';
 import * as timerActions from '../../../features/tracker/actions/timer.actions';
 import * as historyActions from '../../../shared/actions/history.actions';
 import * as platformsActions from '../../../shared/actions/platforms.actions';
+import * as progressActions from '../../../shared/actions/progress.actions';
 import * as userActions from '../../auth/actions/user.actions';
 import * as addPlayingActions from '../../completion/actions/add-playing.actions';
 import * as markCompleteActions from '../../completion/actions/mark-complete.actions';
-import * as progressActions from '../../completion/actions/progress.actions';
 
 import '../../../rxjs-operators';
