+++ src/app/features/completion/effects/progress.effects.ts
@@ -7,9 +7,9 @@ import { Observable } from 'rxjs';
 import { ProgressService } from '../services/progress.service';
 
 import * as appActions from '../../../actions/app.actions';
+import * as progressActions from '../../../shared/actions/progress.actions';
 import * as addPlayingActions from '../actions/add-playing.actions';
 import * as markCompleteActions from '../actions/mark-complete.actions';
-import * as progressActions from '../actions/progress.actions';
 
 @Injectable()
 export class ProgressEffects {
