+++ src/app/features/completion/reducers/root.reducer.spec.ts
@@ -1,9 +1,6 @@
 import { State as AddPlayingState } from './add-playing.reducer';
 import { State as MarkCompleteState } from './mark-complete.reducer';
-import { State as ProgressState } from './progress.reducer';
-import {
-    _selectAddPlayingInfo, _selectCompletedProgress, _selectPlayingProgress, CompletionState, State
-} from './root.reducer';
+import { _selectAddPlayingInfo, CompletionState, State } from './root.reducer';
 
 describe('Completion Root Reducer', () => {
   describe('AddPlaying State Selectors', () => {
@@ -16,7 +13,6 @@ describe('Completion Root Reducer', () => {
         };
         const completionState: CompletionState = {
           addPlaying,
-          progress: initialProgressState,
           markComplete: initialMarkCompleteState
         };
         const state: State = { completion: completionState };
@@ -27,175 +23,7 @@ describe('Completion Root Reducer', () => {
       });
     });
   });
-
-  describe('Progress State Selectors', () => {
-    describe('_selectPlayingProgress', () => {
-      it('Should return an empty list when there are no items', () => {
-        const completionState: CompletionState = {
-          addPlaying: initialAddPlayingState,
-          progress: initialProgressState,
-          markComplete: initialMarkCompleteState
-        };
-        const state: State = { completion: completionState };
-
-        const result = _selectPlayingProgress(state);
-
-        expect(result.length).toBe(0);
-      });
-
-      it('Should return an empty list when all have end entry ids', () => {
-        const progress: ProgressState = {
-          ids: ['1', '2'],
-          entities: {
-            '1': {
-              id: '1',
-              startEntryId: 'start 1',
-              endEntryId: 'end 1'
-            },
-            '2': {
-              id: '2',
-              startEntryId: 'start 2',
-              endEntryId: 'end 2'
-            }
-          },
-          loading: false
-        };
-        const completionState: CompletionState = {
-          addPlaying: initialAddPlayingState,
-          progress,
-          markComplete: initialMarkCompleteState
-        };
-        const state: State = { completion: completionState };
-
-        const result = _selectPlayingProgress(state);
-
-        expect(result.length).toBe(0);
-      });
-
-      it('Should return correct items', () => {
-        const progress: ProgressState = {
-          ids: ['1', '2'],
-          entities: {
-            '1': {
-              id: '1',
-              startEntryId: 'start 1',
-              endEntryId: 'end 1'
-            },
-            '2': {
-              id: '2',
-              startEntryId: 'start 2',
-              endEntryId: ''
-            }
-          },
-          loading: false
-        };
-        const completionState: CompletionState = {
-          addPlaying: initialAddPlayingState,
-          progress,
-          markComplete: initialMarkCompleteState
-        };
-        const state: State = { completion: completionState };
-
-        const result = _selectPlayingProgress(state);
-
-        expect(result).toEqual([{
-          id: '2',
-          startEntryId: 'start 2',
-          endEntryId: ''
-        }]);
-      });
-    });
-
-    describe('_selectCompletedProgress', () => {
-      it('Should return an empty list when there are no items', () => {
-        const completionState: CompletionState = {
-          addPlaying: initialAddPlayingState,
-          progress: initialProgressState,
-          markComplete: initialMarkCompleteState
-        };
-        const state: State = { completion: completionState };
-
-        const result = _selectCompletedProgress(state);
-
-        expect(result.length).toBe(0);
-      });
-
-      it('Should return an empty list when none have end entry ids', () => {
-        const progress: ProgressState = {
-          ids: ['1', '2'],
-          entities: {
-            '1': {
-              id: '1',
-              startEntryId: 'start 1',
-              endEntryId: ''
-            },
-            '2': {
-              id: '2',
-              startEntryId: 'start 2',
-              endEntryId: ''
-            }
-          },
-          loading: false
-        };
-        const completionState: CompletionState = {
-          addPlaying: initialAddPlayingState,
-          progress,
-          markComplete: initialMarkCompleteState
-        };
-        const state: State = { completion: completionState };
-
-        const result = _selectCompletedProgress(state);
-
-        expect(result.length).toBe(0);
-      });
-
-      it('Should return correct items', () => {
-        const progress: ProgressState = {
-          ids: ['1', '2'],
-          entities: {
-            '1': {
-              id: '1',
-              startEntryId: 'start 1',
-              endEntryId: 'end 1'
-            },
-            '2': {
-              id: '2',
-              startEntryId: 'start 2',
-              endEntryId: ''
-            }
-          },
-          loading: false
-        };
-        const completionState: CompletionState = {
-          addPlaying: initialAddPlayingState,
-          progress,
-          markComplete: initialMarkCompleteState
-        };
-        const state: State = { completion: completionState };
-
-        const result = _selectCompletedProgress(state);
-
-        expect(result).toEqual([{
-          id: '1',
-          startEntryId: 'start 1',
-          endEntryId: 'end 1'
-        }]);
-      });
-    });
-  });
 });
-
-const initialAddPlayingState: AddPlayingState = {
-  game: '',
-  platform: '',
-  startTime: 0
-};
-
-const initialProgressState: ProgressState = {
-  ids: [],
-  entities: {},
-  loading: false
-};
 
 const initialMarkCompleteState: MarkCompleteState = {
   ids: [],
