+++ src/app/shared/actions/progress.actions.spec.ts
