+++ src/app/shared/models/index.ts
@@ -1,2 +1,3 @@
 export * from './history.models';
 export * from './shared.models';
+export * from './progress.models';
