+++ src/app/shared/models/progress.models.ts
@@ -1,0 +1,18 @@
+export interface ProgressItem {
+  id: string;
+  startEntryId: string;
+  endEntryId: string;
+}
+
+export interface CompletedItem {
+  game: string;
+  platform: string;
+  startTime: number;
+  endTime: number;
+  timePlayed: number;
+}
+
+export interface MarkCompletePayload {
+  itemId: string;
+  endEntryId: string;
+}
