+++ src/app/shared/reducers/progress.reducer.spec.ts
