+++ src/app/shared/reducers/progress.reducer.ts
